# internal/httpapi/openapi.yaml
openapi: 3.0.3
info:
  title: Statehouse API
  description: >
    Read-only HTTP API for the statehouse home automation state engine.
    All endpoints except /healthz and /openapi.json require a Bearer JWT
    issued by the configured identity service — either a user access token
    or a `client_credentials` service token.


    Browser clients: cross-origin access is governed by the deployment's
    http.allowed_origins allowlist, and is off until one is configured.
    Every path answers an OPTIONS preflight with 204 ahead of
    authentication, since a browser never attaches credentials to one.
    Responses to an allowlisted origin carry Access-Control-Allow-Origin,
    Timing-Allow-Origin and Vary: Origin, on error responses as well as
    successful ones, so a browser client can read a 401 rather than seeing
    an opaque network failure. GET /openapi.json and GET /healthz are
    exempt from the allowlist and always answer
    Access-Control-Allow-Origin: * — both are unauthenticated, so both are
    already readable by any non-browser client, and a spec viewer on an
    arbitrary origin has to be able to fetch the spec.
  version: "1"
servers:
  - url: __PUBLIC_URL__
security:
  - bearerAuth: []
...
paths:
...
  /state:
    get:
      summary: Full state snapshot
      description: Complete snapshot of house state and all devices, with activity and summary.
      operationId: getState
      tags: [state]
      responses:
        "200":
          description: Current snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SnapshotResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /state/house:
    get:
      summary: House state
      description: Occupancy, activity, and mode dimensions for the whole house.
      operationId: getHouseState
      tags: [state]
      responses:
        "200":
          description: Current house state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HouseResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /state/devices:
    get:
      summary: All devices
      description: Map of all known devices keyed by device ID, with full state.
      operationId: getDevices
      tags: [state]
      responses:
        "200":
          description: Map of device ID to device state
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  $ref: "#/components/schemas/DeviceResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /state/activity:
    get:
      summary: Activity state
      description: Active presence/activity signals and recent activity history.
      operationId: getActivity
      tags: [state]
      responses:
        "200":
          description: Current activity state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivityStateResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /events/recent:
    get:
      summary: Recent event log
      description: Recent canonical and derived events from the in-memory log.
      operationId: getRecentEvents
      tags: [events]
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
          description: Maximum number of events to return
      responses:
        "200":
          description: Array of recent events (most-recent last)
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /config/devices:
    get:
      summary: All device profiles
      description: Resolved runtime configuration (class, strategy, thresholds) for all known devices.
      operationId: getDeviceProfiles
      tags: [config]
      responses:
        "200":
          description: Map of device ID to profile
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  $ref: "#/components/schemas/DeviceProfileResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        A Bearer JWT issued by the configured identity service. Two kinds are
        accepted. A **user access token**, obtained by signing in, must belong
        to an account that is still active. A **service token**, obtained with
        the OAuth 2.0 `client_credentials` grant, is accepted for the whole
        API; a deployment may additionally require a matching `aud`, a
        particular `scope`, or restrict callers to an allowlist of `client_id`s
        (see docs/service-tokens.md). Both kinds get the same read-only access
        — there is no per-endpoint distinction between them.
  responses:
    Unauthorized:
      description: >
        Missing or invalid Bearer JWT. The WWW-Authenticate header carries an
        RFC 6750 `error` and `error_description` saying which — an expired
        token, an audience that does not include this server, or a service
        token where the deployment does not accept them. A request with no
        Authorization header at all gets a bare challenge with no error code.
      headers:
        WWW-Authenticate:
          schema:
            type: string
    Forbidden:
      description: >
        The token is valid but the principal is not permitted: a service token
        whose `client_id` is not on this deployment's allowlist, or which does
        not carry the required scope. The WWW-Authenticate header carries
        `error="insufficient_scope"`, and a `scope` parameter naming what is
        required when a scope is the reason. Retrying with the same token will
        not help.
      headers:
        WWW-Authenticate:
          schema:
            type: string
  schemas:
...
    Availability:
      type: string
      enum: [unknown, online, offline_pending, offline]
      description: Whether the engine is currently in contact with the device.
...
    OccupancyState:
      type: string
      enum: [unknown, empty, occupied]
      description: Whole-house occupancy inference.
    HouseActivityState:
      type: string
      enum: [unknown, idle, quiet, active, busy]
      description: Whole-house activity level inference.
...
    HouseOccupancyResponse:
      type: object
      required: [state, confidence, last_changed]
      properties:
        state:
          $ref: "#/components/schemas/OccupancyState"
        confidence:
          type: number
          format: double
        last_changed:
          type: string
          format: date-time
          nullable: true
        last_changed_ago:
          type: integer
          description: Seconds since last state change.
    HouseActivityResponse:
      type: object
      required: [state, confidence, last_changed]
      properties:
        state:
          $ref: "#/components/schemas/HouseActivityState"
        confidence:
          type: number
          format: double
        last_changed:
          type: string
          format: date-time
          nullable: true
        last_changed_ago:
          type: integer
          description: Seconds since last state change.
    HouseModeResponse:
      type: object
      required: [state, confidence, last_changed]
      properties:
        state:
          $ref: "#/components/schemas/ModeState"
        confidence:
          type: number
          format: double
        last_changed:
          type: string
          format: date-time
          nullable: true
        last_changed_ago:
          type: integer
          description: Seconds since last state change.
    HouseResponse:
      type: object
      required: [occupancy, activity, mode, active_devices]
      properties:
        occupancy:
          $ref: "#/components/schemas/HouseOccupancyResponse"
        activity:
          $ref: "#/components/schemas/HouseActivityResponse"
        mode:
          $ref: "#/components/schemas/HouseModeResponse"
        active_devices:
          type: array
          items:
            type: string
          description: IDs of devices currently considered active.

    # ── Device ───────────────────────────────────────────────────────────────
    IdentityResponse:
      type: object
      required: [scheme]
      properties:
        scheme:
          type: string
          description: Adapter protocol ("zigbee", "boiler", "ups", etc.)
        primary:
          type: string
          description: Stable adapter-specific identifier (e.g. IEEE address).
        display:
          type: string
          description: Human-readable mutable name from the adapter.
    ActivityResponse:
      type: object
      required: [state, last_changed, last_changed_ago, confidence]
      properties:
        state:
          $ref: "#/components/schemas/DeviceActivityState"
        last_changed:
          type: string
          format: date-time
          nullable: true
        last_changed_ago:
          type: integer
          nullable: true
          description: Seconds since last state change.
        confidence:
          type: number
          format: double
    LatestResponse:
      type: object
      required: [last_seen, last_seen_ago, stale]
      properties:
        power_w:
          type: number
          format: double
          description: Instantaneous power in watts.
        voltage_v:
          type: number
          format: double
          description: Voltage in volts.
        energy_kwh:
          type: number
          format: double
          description: Cumulative energy in kWh.
        temperature_c:
          type: number
          format: double
          description: Temperature in degrees Celsius.
        humidity_pct:
          type: number
          format: double
          description: Relative humidity percentage.
        pressure_hpa:
          type: number
          format: double
          description: Atmospheric pressure in hPa.
        wind_speed_ms:
          type: number
          format: double
          description: Wind speed in m/s.
        wind_dir_deg:
          type: number
          format: double
          description: Wind direction in degrees.
        rainfall_mm:
          type: number
          format: double
          description: Rainfall in mm.
        illuminance_lux:
          type: number
          format: double
          description: Illuminance in lux.
        uv_index:
          type: number
          format: double
          description: UV index.
        battery_runtime_mins:
          type: number
          format: double
          description: UPS estimated battery runtime in minutes.
        on_battery:
          type: boolean
          description: UPS is running on battery.
        low_battery:
          type: boolean
          description: UPS battery is low.
        smoke:
          type: boolean
          description: Smoke/heat detector primary alarm is active (latched while detecting).
        tamper:
          type: boolean
          description: Smoke/heat detector tamper switch is tripped.
        battery_pct:
          type: number
          format: double
          description: Battery charge percentage.
        linkquality:
          type: integer
          description: Zigbee link quality indicator (0–255).
        rssi_dbm:
          type: integer
          description: Wi-Fi RSSI in dBm.
        last_seen:
          type: string
          format: date-time
          nullable: true
          description: Timestamp of the most recent telemetry message.
        last_seen_ago:
          type: integer
          nullable: true
          description: Seconds since last telemetry message.
        stale:
          type: boolean
          description: True when last_seen exceeds the class staleness threshold.
...
    LifetimeResponse:
      type: object
      description: >-
        All-time per-device aggregates accumulated in memory since process
        start (not persisted). Each field is present only once the device has
        reported that measurement at least once.
      properties:
        max_power_w:
          allOf:
            - $ref: "#/components/schemas/Extremum"
          description: Highest instantaneous power draw observed, in watts.
        min_temperature_c:
          allOf:
            - $ref: "#/components/schemas/Extremum"
          description: Lowest temperature observed, in degrees Celsius.
        max_temperature_c:
          allOf:
            - $ref: "#/components/schemas/Extremum"
          description: Highest temperature observed, in degrees Celsius.
        min_humidity_pct:
          allOf:
            - $ref: "#/components/schemas/Extremum"
          description: Lowest relative humidity observed, as a percentage.
        max_humidity_pct:
          allOf:
            - $ref: "#/components/schemas/Extremum"
          description: Highest relative humidity observed, as a percentage.
...
    CycleResponse:
      type: object
      required: [type, active, started_at, duration_seconds, energy]
      properties:
        type:
          type: string
          enum: [appliance_cycle, compressor_cycle, binary_cycle, unknown]
        active:
          type: boolean
          description: True while the cycle is in progress.
        started_at:
          type: string
          format: date-time
          nullable: true
        finished_at:
          type: string
          format: date-time
          description: Present once the cycle has ended.
        duration_seconds:
          type: integer
          format: int64
          description: Elapsed duration in seconds (growing while active).
        energy:
          $ref: "#/components/schemas/CycleEnergyResponse"
    DeviceResponse:
      type: object
      required: [id, class, availability, activity, latest, warnings]
      properties:
        id:
          type: string
          description: Stable device identifier.
        display_name:
          type: string
          description: Human-readable name override from config.
        class:
          type: string
          description: Device class (e.g. "short_burst_power_device", "continuous").
        room:
          type: string
          description: >
            Floorplan room id this device sits in, e.g. "groundfloor.kitchen".
        floor:
          type: string
          description: >
            Floor id as DECLARED by the devices namespace, e.g. "groundfloor". Absent
            when the namespace declares none — never derived from the room id's
            "<floor>.<slug>" shape, since the floorplan owns that fact.
        covers:
          type: string
          description: >
            What this device's readings describe when that is not its own room:
            "house", or another room id. Absent means it covers the room it is in.
        location:
          type: string
          deprecated: true
          description: >
            DEPRECATED alias for room, carrying the same value. Removed once the
            devices namespace is published and consumers have migrated.
        identity:
          allOf:
            - $ref: "#/components/schemas/IdentityResponse"
          description: Protocol-agnostic identity. Present in /state/devices/{id}, absent in /state/devices.
        availability:
          $ref: "#/components/schemas/Availability"
        activity:
          $ref: "#/components/schemas/ActivityResponse"
        latest:
          $ref: "#/components/schemas/LatestResponse"
        lifetime:
          allOf:
            - $ref: "#/components/schemas/LifetimeResponse"
          description: All-time aggregates. Present once the device reports a tracked measurement.
        cycle:
          allOf:
            - $ref: "#/components/schemas/CycleResponse"
          description: Present for cycle-tracking device classes.
        unclassified:
          type: boolean
          description: True when the device has no matching class config.
        warnings:
          type: array
          items:
            type: string
            enum: [stale_device, cycle_divergence, stale_counter]

    # ── Snapshot ─────────────────────────────────────────────────────────────
    SummaryResponse:
      type: object
      required: [device_count, online_count, active_count, warning_count]
      properties:
        device_count:
          type: integer
        online_count:
          type: integer
        active_count:
          type: integer
          description: Devices currently in an active state (excludes continuous devices).
        warning_count:
          type: integer
          description: Devices with at least one active warning.
    SnapshotResponse:
      type: object
      required: [schema_version, generated_at, summary, house, devices, activity]
      properties:
        schema_version:
          type: string
          example: net.swee.statehouse.snapshot.v1
        site:
          type: string
          example: home
          description: >
            Id of the property this instance reports for, matching an id in the
            `sites` namespace and the `site` tag written on every Influx point.
            Lets a consumer talking to more than one statehouse tell them apart.
            Omitted only when unconfigured, which the service refuses to start in.
        generated_at:
          type: string
          format: date-time
        started_at:
          type: string
          format: date-time
          description: Server start time. Present when uptime data is available.
        started_ago:
          type: integer
          description: Seconds since server start.
        summary:
          $ref: "#/components/schemas/SummaryResponse"
        house:
          $ref: "#/components/schemas/HouseResponse"
        devices:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/DeviceResponse"
        activity:
          $ref: "#/components/schemas/ActivityStateResponse"

    # ── Activity ─────────────────────────────────────────────────────────────
...
    ActivityStateResponse:
      type: object
      required: [generated_at, signals, recent]
      properties:
        generated_at:
          type: string
          format: date-time
        signals:
          type: array
          items:
            $ref: "#/components/schemas/SignalResponse"
          description: Currently active signals.
        recent:
          type: array
          items:
            $ref: "#/components/schemas/ActivityRecordResponse"
          description: Recent completed or in-progress activity records.

    # ── Config ───────────────────────────────────────────────────────────────
...
    DeviceProfileResponse:
      type: object
      required: [class, energy_strategy, resolution]
      properties:
        class:
          type: string
        energy_strategy:
          type: string
          description: Which energy reading to prefer (e.g. "reported", "integrated", "best").
        resolution:
          type: string
          description: Sampling resolution label (e.g. "high", "standard").
        display_name:
          type: string
        room:
          type: string
          description: >
            Floorplan room id this device sits in, e.g. "groundfloor.kitchen". Same
            value and same meaning as DeviceResponse.room.
        floor:
          type: string
          description: >
            Floor id as DECLARED by the devices namespace, e.g. "groundfloor". Absent
            when the namespace declares none — never derived from the room id's
            "<floor>.<slug>" shape, since the floorplan owns that fact.
        covers:
          type: string
          description: >
            What this device's readings describe when that is not its own room:
            "house", or another room id. Absent means it covers the room it is in.
        location:
          type: string
          deprecated: true
          description: >
            DEPRECATED alias for room, carrying the same value. Removed once the
            devices namespace is published and consumers have migrated.
        thresholds:
          $ref: "#/components/schemas/ThresholdsResponse"

    # ── System ───────────────────────────────────────────────────────────────
...